refactor(website): separate account creation and course purchase code
