+++ services/website/src/lib/components/purchase/gumroad/index.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   // import type { Writable } from "svelte/store";
 
-  import { getContext, setContext } from "svelte";
+  import { setContext } from "svelte";
   import { writable } from "svelte/store";
   import {
     contextKeyPurchaseGumroadWizardActiveStep,
